requirements/base.txt
# Core dependencies for CodeRefactor

# HTTP client for Claude API
httpx>=0.23.0

# Web framework for the web interface
Flask>=2.2.0
Flask-WTF>=1.0.0
Flask-Cors>=3.0.10

# Frontend components
python-socketio>=5.7.0

# Code analysis tools
pylint>=2.15.0
mypy>=0.990
flake8>=6.0.0
bandit>=1.7.4
black>=23.1.0
isort>=5.11.0
autopep8>=2.0.0

# Python language utilities
typing-extensions>=4.4.0
dataclasses>=0.8; python_version < '3.7'

# Configuration management
PyYAML>=6.0

# CLI tools
click>=8.1.3
rich>=13.0.0
colorama>=0.4.6

# File handling
pathlib>=1.0.1; python_version < '3.6'

# Performance monitoring
psutil>=5.9.0

# Monaco editor integration
pymonaco>=0.1.0

# .NET integration (optional)
pythonnet>=3.0.0; platform_system == 'Windows' or platform_system == 'Linux'

# Utilities
pydantic>=1.10.0
tqdm>=4.64.0
python-dotenv>=0.21.0